Make the hidden method tests parameterizable
