+++ gems/sorbet/test/hidden-method-finder/simple/sorbet/config
