+++ gems/sorbet/test/hidden-method-finder/simple/src/simple.rb
